Refactored AttachmentDownloadListener to be reusable
