+++ app/src/main/res/layout/activity_nav.xml
@@ -5,7 +5,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NavActivity">
+    tools:activity=".NavActivity">
 
     <ProgressBar
         android:id="@+id/nav_activity_progressbar"
