+++ app/src/main/res/layout/fragment_course_assignments.xml
@@ -4,7 +4,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.development.sakaiclientandroid.fragments.assignments.SiteAssignmentsFragment">
+    tools:activity="com.example.development.sakaiclientandroid.fragments.assignments.SiteAssignmentsFragment">
 
     <!-- View pager for swiping through the assignments -->
     <android.support.v4.view.ViewPager
