+++ app/src/main/res/layout/fragment_single_assignment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.development.sakaiclientandroid.fragments.assignments.SingleAssignmentFragment">
+    tools:activity="com.example.development.sakaiclientandroid.fragments.assignments.SingleAssignmentFragment">
 
     <android.support.v7.widget.CardView
         android:layout_width="match_parent"
